requirements.txt
pyteal==0.8.0
py_algorand_sdk==1.6.0
PyYAML==5.4.1